repair: extract repair_meta
